Added some debugging statements about the mapping.
Did a careful check to make sure that all of the fields that were in the original
mapping are in the AIRR_Mapping file. Annotated each field in the code and commented
it out as it is no longer needed because of the mapping.

Tested the code on the simple example file, seems to work. More testing required.