take care of case float number in metadata
integer numbers are sometimes parsed by python as floats with a zero after the decimal point while API reads them as integers - I added a step to check whether this is happening and to not report it. It will report however if after the decimal there is a value other than 0.