Refactor to ensure an annotation tool specific ID is maintained through the objects (Cell/Clone/Rearrangement).
We want this for data provenance - so we can go back to the files
from the annotation tools and look for oddities when we find them
in the repository. Without this field we can not trace back from
an entity (Cell/Clone/Rearrangement) in the DB to data in a file
on disk.